src/requirements.txt
Django==3.1
coverage==4.2
django-cors-headers==1.3.1
django-filter==2.1.0
django-multiselectfield==0.1.4
djangorestframework==3.11.1
factory-boy==2.12.0
freezegun==0.3.8
psycopg2-binary==2.8.2
pytest-bdd==2.18.1
pytest-capturelog==0.7
pytest-cov==2.3.1
pytest-cover==3.0.0
pytest-datafiles==1.0
pytest-django==3.1.2
pytest-factoryboy==1.3.0
pytest==3.0.2
python-dateutil==2.5.3
requests-mock==1.1.0
requests==2.20.0
celery==4.2.0
django-celery-results==1.0.1
django-debug-toolbar==1.11
django-redis==4.10.0
django-json-widget==0.2.0
django-postgres-extensions==0.9.3
py-kor>=0.0.3
translate>=3.5.0